refactor(arch): Merge fetcher.py and application_utils.py into background.py
