--- app/background_processes/__init__.py
@@ -1,1 +1,0 @@
-from app.background_processes import fetcher  # noqa: F401 #isort:skip
